# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: IOT Tofu Server
  description: 'IOT Tofu Server vytvořený v rámci kurzu Java SpringBootu'
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: https://petstore.swagger.io/v2
  - url: http://localhost:8080
paths:
  /identification:
    get:
      summary: Načtení identifikace IOT serveru
      operationId: getIdentification
      responses:
        200:
          description: OK response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identification'
        500:
          description: Something went horribly terribly wrong
  /meteorological-data:
    post:
      summary: Save meteorogical data
      operationId: saveMeteorologicalData
      requestBody:
        $ref: '#/components/requestBodies/MeteorologicalDataRequestBody'
      responses:
        200:
          description: OK response
        400:
          description: Invalid input
    get:
      summary: Get all saved meteorogical data
      operationId: getMeteorologicalData
      responses:
        200:
          description: OK response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MeteorologicalData'
components:
  schemas:
    Address:
      required:
        - city
      type: object
      properties:
        street:
          type: string
        number:
          type: string
        city:
          type: string
    Identification:
      required:
        - serverName
        - address
        - owner
      type: object
      properties:
        serverName:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        owner:
          $ref: '#/components/schemas/Owner'
    MeteorologicalData:
      type: object
      properties:
        temperature:
          type: number
          format: float
        humidity:
          type: number
          format: float
    Owner:
      required:
        - firstName
        - lastName
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
  requestBodies:
    MeteorologicalDataRequestBody:
      description: Meteorogical data to save
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MeteorologicalData'
            required: true